refactor cmd/skywire-cli/commands/tp/tp.go into multiple files
